Breaks off some form tabs into components
